Moved mcap class
Renamed class Mcap to DepthAiMcap and moved to libraries/depthai_mcap.py
Changed class, so that all streams are saved to a single file.
