+++ gen2-record-replay/libraries/depthai_record.py
@@ -9,6 +9,7 @@ from multiprocessing import Queue
 from pathlib import Path
 from threading import Thread
 
+from depthai_mcap import DepthAiMcap
 import depthai as dai
 import numpy as np
 from mcap.mcap0.writer import Writer
@@ -43,7 +44,8 @@ class Record:
 
     def run(self):
         files = {}
-        mcaps = {}
+        if self.mcap:
+            mcap = DepthAiMcap(str(self.path / f"recordings"))
 
         def create_video_file(name):
             if name == 'depth':  # or (name=='color' and 'depth' in self.save):
@@ -56,10 +58,6 @@ class Record:
             # else : fourcc = "GREY" #Simple, single Y plane for monochrome images.
             # files[name] = VideoWriter(str(path / f"{name}.avi"), VideoWriter_fourcc(*fourcc), fps, sizes[name],
             # isColor=name=="color")
-
-        def create_mcap_file(name):
-            mcaps[name] = Mcap(str(self.path / f"{name}"))
-            mcaps[name].imageInit()
 
         while True:
             try:
@@ -78,10 +76,10 @@ class Record:
                         # frames[name].tofile(files[name])
 
                     else:
-                        if name not in mcaps:  # File wasn't created yet
-                            create_mcap_file(name)
+                        if name not in mcap.channels:  # File wasn't created yet
+                            mcap.imageInit(name)
 
-                        mcaps[name].imageSave(frames[name])
+                        mcap.imageSave(frames[name], name)
 
             except KeyboardInterrupt:
                 break
@@ -90,8 +88,7 @@ class Record:
             for name in files:
                 files[name].close()
         else:
-            for name in mcaps:
-                mcaps[name].close()
+            mcap.close()
         print('Exiting store frame thread')
 
     def start(self):
@@ -269,194 +266,3 @@ class Record:
         self.nodes = nodes
         self.pipeline = pipeline
         return pipeline, nodes
-
-
-class Mcap:
-
-    # when initialising send in path (folder most likely: "./recordings/-name-") without .mcap at the end
-    def __init__(self, path):
-        self.fileName = path + ".mcap"
-        self.stream = open(self.fileName, "wb")
-        self.writer = Writer(self.stream)
-        self.writer.start(profile="x-custom", library="my-writer-v1")
-        self.channel_id = None
-
-    def imageInit(self):
-        # create schema for the type of message that will be sent over to foxglove
-        # for more details on how the schema must look like visit:
-        # http://docs.ros.org/en/noetic/api/sensor_msgs/html/index-msg.html
-        schema_id = self.writer.register_schema(
-            name="ros.sensor_msgs.CompressedImage",
-            encoding="jsonschema",
-            data=json.dumps(
-                {
-                    "type": "object",
-                    "properties": {
-                        "header": {
-                            "type": "object",
-                            "properties": {
-                                "stamp": {
-                                    "type": "object",
-                                    "properties": {
-                                        "sec": {"type": "integer"},
-                                        "nsec": {"type": "integer"},
-                                    },
-                                },
-                            },
-                        },
-                        "format": {"type": "string"},
-                        "data": {"type": "string", "contentEncoding": "base64"},
-                    },
-                },
-            ).encode()
-        )
-
-        # create and register channel
-        self.channel_id = self.writer.register_channel(
-            schema_id=schema_id,
-            topic="image",
-            message_encoding="json",
-        )
-
-    def pointCloudInit(self):
-        # create schema for the type of message that will be sent over to foxglove
-        # for more details on how the schema must look like visit:
-        # http://docs.ros.org/en/noetic/api/sensor_msgs/html/index-msg.html
-        schema_id = self.writer.register_schema(
-            "ros.sensor_msgs.PointCloud2",
-            "jsonschema",
-            data=json.dumps(
-                {
-                    "type": "object",
-                    "properties": {
-                        "header": {
-                            "type": "object",
-                            "properties": {
-                                "seq": {"type": "integer"},
-                                "stamp": {
-                                    "type": "object",
-                                    "properties": {
-                                        "sec": {"type": "integer"},
-                                        "nsec": {"type": "integer"},
-                                    },
-                                },
-                                "frame_id": {"type": "string"}
-                            },
-                        },
-                        "height": {"type": "integer"},
-                        "width": {"type": "integer"},
-                        "fields": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "properties": {
-                                    "name": {"type": "string"},
-                                    "offset": {"type": "integer"},
-                                    "datatype": {"type": "integer"},
-                                    "count": {"type": "integer"}
-                                }
-                            },
-                        },
-                        "is_bigendian": {"type": "boolean"},
-                        "point_step": {"type": "integer"},
-                        "row_step": {"type": "integer"},
-                        "data": {"type": "string", "contentEncoding": "base64"},
-                        "is_dense": {"type": "boolean"}
-                    },
-                },
-            ).encode("utf8")
-        )
-
-        # create and register channel
-        self.channel_id = self.writer.register_channel(
-            schema_id=schema_id,
-            topic="pointClouds",
-            message_encoding="json",
-        )
-
-    # send in image read with "cv2.getCvFrame"
-    def imageSave(self, img):
-        # convert cv2 image to .jpg format
-        # is_success, im_buf_arr = cv2.imencode(".jpg", img)
-
-        # read from .jpeg format to buffer of bytes
-        # byte_im = im_buf_arr.tobytes()
-        byte_im = img.tobytes()
-
-        # data must be encoded in base64
-        data = base64.b64encode(byte_im).decode("ascii")
-
-        tmpTime = time.time_ns()
-        sec = math.trunc(tmpTime / 1e9)
-        nsec = tmpTime - sec
-
-        self.writer.add_message(
-            channel_id=self.channel_id,
-            log_time=tmpTime,
-            data=json.dumps(
-                {
-                    "header": {"stamp": {"sec": sec, "nsec": nsec}},
-                    "format": "jpeg",
-                    "data": data,
-                }
-            ).encode("utf8"),
-            publish_time=tmpTime,
-        )
-
-    # send in point cloud object read with
-    # "o3d.io.read_point_cloud" or
-    # "o3d.geometry.PointCloud.create_from_depth_image"
-    # seq is just a sequence number that will be incremented in main  program (int from 0 to number at end of recording)
-    def pointCloudSave(self, pcd, seq):
-        points = np.asarray(pcd.points)
-
-        # points must be read to a buffer and then encoded with base64
-        buf = bytes()
-        for point in points:
-            buf += struct.pack('f', float(point[0]))
-            buf += struct.pack('f', float(point[1]))
-            buf += struct.pack('f', float(point[2]))
-
-        data = base64.b64encode(buf).decode("ascii")
-
-        tmpTime = time.time_ns()
-        sec = math.trunc(tmpTime / 1e9)
-        nsec = tmpTime - sec
-
-        self.writer.add_message(
-            channel_id=self.channel_id,
-            log_time=time.time_ns(),
-            data=json.dumps(
-                {
-                    "header": {
-                        "seq": seq,
-                        "stamp": {"sec": sec, "nsec": nsec},
-                        "frame_id": "front"
-                    },
-                    "height": 1,
-                    "width": len(pcd),
-                    "fields": [{"name": "x", "offset": 0, "datatype": 7, "count": 1},
-                               {"name": "y", "offset": 4, "datatype": 7, "count": 1},
-                               {"name": "z", "offset": 8, "datatype": 7, "count": 1}],
-                    "is_bigendian": False,
-                    "point_step": 12,
-                    "row_step": 12 * len(pcd),
-                    "data": data,
-                    "is_dense": True
-                }
-            ).encode("utf8"),
-            publish_time=time.time_ns(),
-        )
-
-    def imuInit(self):
-        # TODO create imu support
-        return
-
-    def imuSave(self):
-        # TODO create imu support
-        return
-
-    def close(self):
-        # end writer and close opened file
-        self.writer.finish()
-        self.stream.close()
